CategoriesModelTest: changes mage to CategoriesModel and tested
