+++ app/src/main/java/fr/free/nrw/commons/category/CategoriesModel.java
@@ -123,8 +123,9 @@ public class CategoriesModel{
         }
 
         //otherwise, search API for matching categories
+        //term passed as lower case to make search case-insensitive(taking only lower case for everything)
         return categoryClient
-                .searchCategoriesForPrefix(term, SEARCH_CATS_LIMIT)
+                .searchCategoriesForPrefix(term.toLowerCase(), SEARCH_CATS_LIMIT)
                 .map(name -> new CategoryItem(name, false));
     }
 
@@ -188,7 +189,7 @@ public class CategoriesModel{
      * @return
      */
     private Observable<CategoryItem> getTitleCategories(String title) {
-        return categoryClient.searchCategories(title, SEARCH_CATS_LIMIT)
+        return categoryClient.searchCategories(title.toLowerCase(), SEARCH_CATS_LIMIT)
                 .map(name -> new CategoryItem(name, false));
     }
