+++ app/src/main/java/fr/free/nrw/commons/category/CategoryClient.java
@@ -30,7 +30,6 @@ public class CategoryClient {
 
     /**
      * Searches for categories containing the specified string.
-     * filter passed as lower case to make search case-insensitive(taking only lower case for everything)
      *
      * @param filter    The string to be searched
      * @param itemLimit How many results are returned
@@ -38,7 +37,7 @@ public class CategoryClient {
      * @return
      */
     public Observable<String> searchCategories(String filter, int itemLimit, int offset) {
-        return responseToCategoryName(CategoryInterface.searchCategories(filter.toLowerCase(), itemLimit, offset));
+        return responseToCategoryName(CategoryInterface.searchCategories(filter, itemLimit, offset));
 
     }
 
@@ -56,7 +55,6 @@ public class CategoryClient {
 
     /**
      * Searches for categories starting with the specified string.
-     * prefix passed as lower case to make search case-insensitive
      * 
      * @param prefix    The prefix to be searched
      * @param itemLimit How many results are returned
@@ -64,7 +62,7 @@ public class CategoryClient {
      * @return
      */
     public Observable<String> searchCategoriesForPrefix(String prefix, int itemLimit, int offset) {
-        return responseToCategoryName(CategoryInterface.searchCategoriesForPrefix(prefix.toLowerCase(), itemLimit, offset));
+        return responseToCategoryName(CategoryInterface.searchCategoriesForPrefix(prefix, itemLimit, offset));
     }
 
     /**
